task_service/requirements.txt
Django==4.0.4
psycopg2-binary>=2.8
pytest-django==4.5.2
snowflake-id==0.0.2
requests-oauthlib==1.3.1
PyJWT==2.3.0